Rider suggestions (#4778)
* more tests

* more suggestions

* more cleanup

* more imporvements

* fix compile error

+++ Project/Assets/ML-Agents/Examples/FoodCollector/Scripts/FoodCollectorAgent.cs
@@ -1,4 +1,3 @@
-using System;
 using UnityEngine;
 using Unity.MLAgents;
 using Unity.MLAgents.Actuators;
@@ -106,7 +105,7 @@ public class FoodCollectorAgent : Agent
             dirToGo += transform.right * right;
             rotateDir = -transform.up * rotate;
 
-            var shootCommand = (int)discreteActions[0] > 0;
+            var shootCommand = discreteActions[0] > 0;
             if (shootCommand)
             {
                 m_Shoot = true;
